Calculate scores for all permutations
Scores are calculated for each permutation based on how many turns
the AI makes in the game (shorter games are better outcomes). An array
of data objects with a unique guid and a score is created. If a game
finishes in less than 9 moves, permutations that could not occur as a
result do not have score objects created.