Recursive minmax logic implemented
Game AI will recursively look at its array of possible outcomes and
evaluate the optimal min/max choices. Sequence and functionality works
as intended but there is an issue with the logic. The issue appears most
clearly when there is an option to block an opponents winning move and
another square is chosen.