Wake up heroku
Application makes a GET request to the API server on document.ready.
This should give the user time to sign up/in while the heroku instance
wakes up.

Added .fail() to the GET request to trap any responses and prevent
console errors since the default route for the heroku server will
return an authentication or 404 error